Fix: Supervise Rooftop 수정
